Trait pallet_robonomics_rws::pallet::Config

pub trait Config: Config {
    type Call: Parameter + UnfilteredDispatchable<Origin = Self::Origin> + GetDispatchInfo;
    type Time: Time<Moment = Self::Moment>;
    type Moment: Parameter + AtLeast32Bit + Into<Weight>;
    type AuctionIndex: Parameter + AtLeast32Bit + Default;
    type AuctionCurrency: ReservableCurrency<Self::AccountId>;
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type ReferenceCallWeight: Get<Weight>;
    type WeightLimit: Get<Weight>;
    type AuctionDuration: Get<Self::BlockNumber>;
    type AuctionCost: Get<<<Self as Config>::AuctionCurrency as Currency<<Self as Config>::AccountId>>::Balance>;
    type MinimalBid: Get<<<Self as Config>::AuctionCurrency as Currency<<Self as Config>::AccountId>>::Balance>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Call: Parameter + UnfilteredDispatchable<Origin = Self::Origin> + GetDispatchInfo

Call subscription method.

type Time: Time<Moment = Self::Moment>

Current time source.

type Moment: Parameter + AtLeast32Bit + Into<Weight>

Time should be aligned to weights for TPS calculations.

type AuctionIndex: Parameter + AtLeast32Bit + Default

The auction index value.

type AuctionCurrency: ReservableCurrency<Self::AccountId>

The auction bid currency.

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

The overarching event type.

type ReferenceCallWeight: Get<Weight>

Reference call weight, general transaction consumes this weight.

type WeightLimit: Get<Weight>

Subscription weight accumulator limit.

type AuctionDuration: Get<Self::BlockNumber>

Subscription auction duration in blocks.

type AuctionCost: Get<<<Self as Config>::AuctionCurrency as Currency<<Self as Config>::AccountId>>::Balance>

How much token should be bonded to launch new auction.

type MinimalBid: Get<<<Self as Config>::AuctionCurrency as Currency<<Self as Config>::AccountId>>::Balance>

Minimal auction bid.